requirements.txt
asgiref==3.7.2
Django==4.2
djangorestframework==3.14.0
idna==3.6
jwcrypto==1.5.1
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
